refactor: move prototype files into src
