+++ src/index.html
@@ -19,6 +19,6 @@
         </div>
       </div>
     </div>
-    <script type="module" src="/src/main.ts"></script>
+    <script type="module" src="/game.ts"></script>
   </body>
 </html>
