+++ vite.config.ts
@@ -1,0 +1,6 @@
+import { defineConfig } from 'vite';
+
+export default defineConfig({
+  root: 'src',
+  publicDir: '../public'
+});
